flask_api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.3.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
priority==2.0.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-environ==0.4.54
soupsieve==2.4.1
toml==0.10.2
tomli==2.0.1
Werkzeug==2.3.0
wsproto==1.2.0
yarl==1.9.2
